Minimal fixes for C++-compliance.
Allow palettes with >256 entries.

+++ timer.cpp
@@ -16,6 +16,6 @@ long timer_get_elapsed(Timer* t)
     if (t->tv_start.tv_sec == t->tv_end.tv_sec)
         return t->tv_end.tv_usec - t->tv_start.tv_usec;
     else
-        return (t->tv_end.tv_sec - t->tv_start.tv_sec) *
-            1e6 + (t->tv_end.tv_usec - t->tv_start.tv_usec);
+        return long((t->tv_end.tv_sec - t->tv_start.tv_sec) *
+            1e6 + (t->tv_end.tv_usec - t->tv_start.tv_usec));
 }
